reorganizing spectral model
