--- chemprop/models/v2/layers.py
@@ -1,5 +1,0 @@
-from torch import Tensor, nn
-
-class Exp(nn.Module):
-    def forward(self, x: Tensor):
-        return x.exp()
